API Added a SimpleMemberList management
Added ability to use a ListOfMembers interface for defining a set of members
that can be granted permission to an item. This simplifies the need to use
Group objects, which can in some use-cases pollute the Group section of the
CMS, and provide overheads (eg inheritance) that isn't needed.

The SimpleMemberList provides a very basic implementation of such a list.